Fix code sample
First of all I love this library, it's such great work!

I noticed that several comments on the /r/programming and HackerNews got really hung up on the sample, despite all the other great work in the library (internet commentators eh!!). 

So this P.R makes the code sample match the latest version of your code. I've left the formatting as the tool outputs it by via calling `ToString()` on the `SyntaxNode`